+++ src/ai/backend/manager/models/resource_policy.py
@@ -13,15 +13,6 @@ from sqlalchemy.orm import relationship, selectinload
 from ai.backend.common.types import DefaultForUnspecified
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.models.utils import execute_with_retry
-from ai.backend.manager.services.keypair_resource_policies.actions.create_keypair_resource_policy import (
-    CreateKeyPairResourcePolicyAction,
-)
-from ai.backend.manager.services.keypair_resource_policies.actions.delete_keypair_resource_policy import (
-    DeleteKeyPairResourcePolicyAction,
-)
-from ai.backend.manager.services.keypair_resource_policies.actions.modify_keypair_resource_policy import (
-    ModifyKeyPairResourcePolicyAction,
-)
 
 from .base import (
     Base,
@@ -393,9 +384,6 @@ class CreateKeyPairResourcePolicy(graphene.Mutation):
         return CreateKeyPairResourcePolicy(
             ok=True,
             msg="",
-            resource_policy=KeyPairResourcePolicy.from_row(
-                graph_ctx, result.keypair_resource_policy
-            ),
         )
 
 
@@ -417,17 +405,18 @@ class ModifyKeyPairResourcePolicy(graphene.Mutation):
         name: str,
         props: ModifyKeyPairResourcePolicyInput,
     ) -> ModifyKeyPairResourcePolicy:
+        from ai.backend.manager.services.keypair_resource_policies.actions.modify_keypair_resource_policy import (
+            ModifyKeyPairResourcePolicyAction,
+        )
+
         graph_ctx: GraphQueryContext = info.context
-        result = await graph_ctx.processors.keypair_resource_policy_service.modify_keypair_resource_policy.wait_for_complete(
+        await graph_ctx.processors.keypair_resource_policy_service.modify_keypair_resource_policy.wait_for_complete(
             ModifyKeyPairResourcePolicyAction(name, props)
         )
 
         return ModifyKeyPairResourcePolicy(
             ok=True,
             msg="",
-            resource_policy=KeyPairResourcePolicy.from_row(
-                graph_ctx, result.keypair_resource_policy
-            ),
         )
 
 
@@ -447,16 +436,17 @@ class DeleteKeyPairResourcePolicy(graphene.Mutation):
         info: graphene.ResolveInfo,
         name: str,
     ) -> DeleteKeyPairResourcePolicy:
+        from ai.backend.manager.services.keypair_resource_policies.actions.delete_keypair_resource_policy import (
+            DeleteKeyPairResourcePolicyAction,
+        )
+
         graph_ctx: GraphQueryContext = info.context
-        result = await graph_ctx.processors.keypair_resource_policy_service.delete_keypair_resource_policy.wait_for_complete(
+        await graph_ctx.processors.keypair_resource_policy_service.delete_keypair_resource_policy.wait_for_complete(
             DeleteKeyPairResourcePolicyAction(name)
         )
         return DeleteKeyPairResourcePolicy(
             ok=True,
             msg="",
-            resource_policy=KeyPairResourcePolicy.from_row(
-                graph_ctx, result.keypair_resource_policy
-            ),
         )
